+++ CHANGELOG.md
@@ -1,3 +1,9 @@
+0.1.2
+- Fixed some internal issues.
+- Fixed class link attribute behavior.
+- Added property name to RouteInstance.
+- Added ability to add active class to a parent element.
+
 0.1.1
 - Updated TreborJS to the latest version.